hr/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
Django==3.0.5
django-bootstrap4==1.1.1
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
pytz==2020.1
redis==3.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
